Primeros cambios

+++ backend-AMPA/database/migrations/2023_03_14_195128_create_cursos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('cursos', function (Blueprint $table) {
             $table->id();
             $table->string('nombre');
-            $table->integer('precio_cuota');
+            $table->float('precio_cuota');
             $table->date('fecha_inicio');
             $table->date('fecha_fin');
             $table->timestamps();
