+++ backend-AMPA/database/migrations/2023_03_14_195205_create_eventos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('titulo');
             $table->string('descripcion');
             $table->integer('capacidad');
-            $table->double('precio');
+            $table->float('precio');
             $table->timestamps();
         });
     }
